+++ FixedPopupPanels/FixedPopupPanels.js
@@ -46,6 +46,13 @@ const fixedPopupPanelsHandlers = {
 		forProp('rtl', false),
 		forKey('left'),
 		(ev, {index}) => (index > 0),
+		(ev) => {
+			if (ev.target.tagName === 'INPUT') {
+				ev.stopPropagation();
+				return false;
+			}
+			return true;
+		},
 		({target}) => (document.querySelector(`section.${css.body}`).contains(target)),
 		({target}) => (getTargetByDirectionFromElement('left', target) === null),
 		forwardCustom('onBack'),